05-data-platforms/my-pipeline/pipeline/assets/ingestion/requirements.txt
# TODO: Add Python dependencies required by the python ingestion asset.
# - Keep this intentionally minimal.
# - Pin versions for reproducibility.
# Docs: https://getbruin.com/docs/bruin/assets/python

numpy<2
pandas
pyarrow<15
requests
python-dateutil
duckdb